python-voice-model-main/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.6
pipecat-ai[groq,silero]==0.0.90
python-socketio==5.11.3